requirements.txt
# ──── Web Scraping ────
curl_cffi>=0.7.1
trafilatura>=1.8.1
beautifulsoup4>=4.12.0
googlenewsdecoder>=0.1.7
requests>=2.31.0

# ──── AI / ML ────
transformers>=4.40.1
torch>=2.2.2
sentence-transformers>=2.7.0
scipy>=1.13.0
numpy>=1.26.0

# ──── Database ────
neo4j>=5.19.0

# ──── NLP Utilities ────
jellyfish>=1.0.0
python-dateutil>=2.9.0

# ──── UI ────
streamlit>=1.33.0
streamlit-agraph>=0.0.45
pandas>=2.2.0

# ──── Orchestration ────
schedule>=1.2.1

# ──── Config ────
python-dotenv>=1.0.0
